Move Umbraco "style_formats" to seperate heading when combining.
Umbraco currently throws all styles in together so if you have several styles they end up in one big list.  In my opinion this is pretty messy so I have setup menus in tinyMceConfig.config.  When the "style_format" objects are combined it messes with the layout so I think if the custom setting for "style_formats" exists then it makes sense to move the Umbraco generated ones to a subheader as you can still add  custom styles to that subheading from config without modifying core code but this way it won't mess with new subheadings you specify.

Here is screenshot to show you what I mean: http://prntscr.com/6jb3a9

Maybe rename "custom" to name of the site or something. 

I think it should pass the RTE property editor to scope so another set of settings can be loaded.  This would let you have multiple RTE with different settings.  

Also I'd like to allow for inline RTE editing so that you can have wrapper HTML around the RTE so it resembles the element on the site.  This also means you do not have to maintain 2 separate css branches as you can set the right inheritance for the section the inline editor is being inserted to but I think a completely new property editor might be best for this.